pyproject.toml
[tool.poetry]
name = "smartlead-python-client"
version = "0.1.2"
description = ""
authors = ["Ohio Imevbore <ohio.imevbore@gmail.com>"]
readme = "README.md"
packages = [
    { include = "smartlead" }
]

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
requests = "^2.31.0"

[tool.poetry.group.test.dependencies]
python-dotenv = "^1.0.0"
pytest = "^7.4.3"
pandas = "^2.1.3"
jupyter = "^1.0.0"
streamlit = "^1.28.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
